requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
emoji==2.8.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==9.0.0
psycopg2-binary==2.9.9
PyMySQL==1.1.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
telegram==0.0.1
typing_extensions==4.9.0
urllib3==2.0.7
Werkzeug==3.0.1
